cli: internal refactor of backend start implementation
